tmpfile() is now used. Ways is now a define.

+++ tp1/src/attack.h
@@ -7,6 +7,6 @@ typedef struct {
     int y;
 } Attack;
 
-long long attackCompar(const void *a, const void *b);
+int attackCompar(const void *a, const void *b);
 
 #endif
